wizard: display warning when focus is lost
Display the warning that the password is stored in plaintext after the
focus of the password input field is lost.

The current behavior of showing the warning after the first character is
entered is ackward and confusing.

It also eliminates the need to debounce the warning when a password is
pasted.

Reported-by: Brad <super1337@posteo.net>
Signed-off-by: Koni Marti <koni.marti@gmail.com>
Acked-by: Robin Jarry <robin@jarry.cc>