export-mbox: better path suggestion and name completion
Change the `:export-mbox` path completion algorithm, so it works the
same as `:import-mbox`: the user can select folders with auto-complete,
the ~ symbol works as home folder and so on.

Move the automatic mbox-file naming into the export function and only
use it if the user-supplied path is an existing folder.

Signed-off-by: Vitaly Ovchinnikov <v@postbox.nz>
Acked-by: Robin Jarry <robin@jarry.cc>