release.sh: fix email encoding
The release emails are sent without any Content-Transfer-Encoding nor
Content-Type headers. This causes non-ASCII characters to be rendered as
garbage by email clients. Here are two examples from the latest release:

	Matï¿½ï¿½j Cepl
	Nojus Gudinaviï¿½ï¿½ius

Everything is UTF-8 locally, state it explicitly in the email. Use 8bit
as transfer encoding since we are not quoting anything.

Add a From: header since some sendmail implementations do not add
a default value when it is missing.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: Moritz Poldrack <moritz@poldrack.dev>