carddav-query: interpret percent escapes in source url
The username and password must be percent encoded if they are set in
carddav-source. Python's parser does not automatically unquote the
different URL elements.

Some carddav servers do not care about this but some do and report
authentication failures when the username is percent encoded.

Make sure to unquote the username and password after reading them from
carddav-source.

Reported-by: Callum Andrew <contact@candrew.net>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Callum Andrew <contact@candrew.net>