contrib: add a script to generate release statistics
Add a new python script, that will generate two types of statistics for
each minor release (and HEAD) from the local repository and create
a graph. First statistics is number of commits since the previous
release, the second statistics is number of lines inserted and deleted
since the previous release.

The placement of the text annotations (tag names) is mildly smart, so
the distribution changes significantly the hard coded numbers might have
to be tweaked.

Requires matplotlib to be installed (developed with 3.7.3). It generates
a png by default, but this can be changed by passing the output file
option with a different file extension as long as matplotlib knows what
to do with it.

Signed-off-by: Bence Ferdinandy <bence@ferdinandy.com>
Acked-by: Robin Jarry <robin@jarry.cc>