binds: remove duplicated and empty bindings
Add keybindings filtering, so if the context binding is defined, the
parent one is removed and is not shown in the list at all.

Also add keybinding removing, so if the context defines an empty
binding, both this and parent one are removed from the list. This way
you can disable and hide specific bindings in contexts.

Changelog-added: Disable parent context bindings by declaring them empty.
Signed-off-by: Vitaly Ovchinnikov <v@ovch.ru>
Acked-by: Robin Jarry <robin@jarry.cc>