compose: fix header navigation after :compose -e
When the terminal is closed with [compose].edit-headers=true, all
headers are deleted and recreated based on the email content.

Since the terminal is not active, adding the first header was working
fine, but the next ones were replacing the single entry on each call of
addEditor().

Fix the broken append() logic.

Reported-by: Inwit <inwit@sindominio.net>
Fixes: c2a4fc7fdfae ("compose: avoid panic when deleting the last header")
Changelog-fixed: Selection of headers in composer after `:compose -e`
 followed by `:edit -E`.
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Inwit <inwit@sindominio.net>