linters: avoid crash when analyzing function call
When encountering a statement such as:

	go functionName()

The identifier is not a local symbol but should be looked up in the
current package. Do not consider that all these statements will refer to
local variables.

The only way to do this is to run a second analyzer that depends on the
first one. Store all unresolved methods and functions into
a indirectCalls struct. Reuse that result in the second analyzer to
resolve the function declarations and check their bodies for the
defer log.PanicHandler() statement.

Fix one newly reported issues.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Reviewed-by: Tim Culverhouse <tim@timculverhouse.com>