cf: fix over quoting of notmuch queries
Currently, :cf thread:\{id:{{.MessageId}}\} is broken because it is
quoted before being interpreted by notmuch. The dynamic folder is
created with this "query" (including the quotes):

	'thread:{id:23627381....}'

Notmuch queries use the xapian syntax and do not follow basic shell
quotes interpretation.

Change :cf only argument to preserve the command line as entered by the
user without any interpretation. When the backend is notmuch, forward
that as the dynamic folder name.

For other backends, interpret shell quoting on the user entry and fail
if it produces more than one argument.

Link: https://xapian.org/docs/queryparser.html
Fixes: 6613d9b555be ("cf: fix unexpected argument on notmuch")
Reported-by: Inwit <inwit@sindominio.net>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Inwit <inwit@sindominio.net>