rmdir: back to current folder upon error
Jump back to the current folder consistently when encountering an error.
This has only been partly implemented. For example, when you try to
remove a notmuch folder, the remove operation will fail because it is
unsupported. However, you would end up in a different directory because
a directoy change is done before trying to remove the current folder. If
this happens make sure you end up in the current directory again.

Fixes: a35d9bab4664 ("rmdir: ensure proper sequence of operations")
Signed-off-by: Koni Marti <koni.marti@gmail.com>
Acked-by: Robin Jarry <robin@jarry.cc>