ui: make viewer tab title configurable via templates
We had account and composer tab title configuration fields already, but
not for viewer. Add tab-title-viewer configuration, which defaults to
Subject if it is not empty and to "(no subject)" when it is empty.

Changelog-added: New `[ui].tab-title-viewer` setting to configure the
 message viewer tab title.
Changelog-changed: Message viewer tab titles will now show `(no subject)`
 if there is no subject in the viewed email.
Signed-off-by: Bence Ferdinandy <bence@ferdinandy.com>
Acked-by: Robin Jarry <robin@jarry.cc>