dirtree: prevent folder selection when archiving
Prevent an unwanted folder selection in the directory tree when
archiving.

A previous fix (commit 0be135a3) introduced a regression which caused
the directory tree to select the folder for which a CreateDirectory
message returned successfully.

Since the :archive command will issue a CreateDirectory message in every
operation, the directory tree will thus erroneously select the target
archive directory.

However, the functionality to select a newly created folder by the
:mkdir command is actually correct and should be kept since this
corresponds to the regular directory list behavior.

Another quirk that is addressed is caused by the rebuilding of the
directory tree when the underyling directories change. In that case, the
ui.DirListCollapse setting would overwrite the user changes. Therefore,
only apply the ui.DirListCollapse setting at the first time of building
the directory tree.

Fixes: 0be135a38186 ("dirtree: fix jumping folders")
Signed-off-by: Koni Marti <koni.marti@gmail.com>
Acked-by: Robin Jarry <robin@jarry.cc>
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>