imap: show parseable messages on error
Currently if there are mails aerc can not properly handle, listing the
mailbox is aborted. This is especially bothersome if the message one is
looking for would be parseable without issue.

Show an error if parsing of a message fails, but handle all other
messages as normal. This leads to those messages showing up as still
loading.

Signed-off-by: Moritz Poldrack <git@moritz.sh>
Acked-by: Robin Jarry <robin@jarry.cc>