notmuch: keep track of all siblings in thread
Consider a thread with a root message and three child messages, each of
which has a single child message, like so:

    root
        child 1
            grandchild 1a
        child 2
            grandchild 2a
        child 3
            grandchild 3a

With the previous implementation, if child 2 (or child 3) is not shown
by the current query, then grandchild 2a (or grandchild 3a) will not be
shown either. Ensure that this bug does not occur; that is, ensure that
children of non-queried non-first-child messages are shown in threads.

A more complex implementation is possible that maintains only a single
loop over the messages, but thinking about it made my head hurt. Use a
second (at times redundant) loop instead for simplicity's sake.

Changelog-fixed: Don't lose child messages of non-queried parents in
 notmuch threads
Signed-off-by: Jason Cox <me@jasoncarloscox.com>
Acked-by: Robin Jarry <robin@jarry.cc>
Tested-by: Inwit <inwit@sindominio.net>