bindings: do not systematically trigger completion
When simulating keystrokes of a binding, the command completion is
disabled momentarily for performance reasons and re-enabled once the
sequence is finished. See commit 055c6dc6604f ("exline: don't draw
completions for keybinds") for more details.

Since commit 0b0095eeadaf ("complete: allow disabling automatic
completion"), it is possible to only rely on explicit keystrokes to
display the completion menu.

With the default settings, if a key sequence contains more than
[ui].completion-min-chars, it should trigger completion after
[ui].completion-delay. But since the completion was disabled when the
keystrokes are input, it does not trigger the completion.

To work around this, an artificial <Tab> keystroke was added at the end
of the sequence to force trigger the completion menu. For more details,
see commit 04869bd2a39a ("aerc: fix popover menu regression").

The workaround that was added, along with commit b3dc63d69c14
("complete: only display popover for more than one choice"), forces the
completion when there is a single choice. Completely ignoring
[ui].completion-min-chars = manual.

Only explicitly trigger the completion if the completion key was seen in
the keystroke sequence or if completion-min-chars is not set to manual.
Use the correct completion key and not hard code Tab.

Fixes: 0b0095eeadaf ("complete: allow disabling automatic completion")
Fixes: https://todo.sr.ht/~rjarry/aerc/210
References: https://todo.sr.ht/~rjarry/aerc/104
Cc: Skejg <grolleman@zoho.com>
Reported-by: Karel Balej <balejk@matfyz.cz>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Karel Balej <balejk@matfyz.cz>
Tested-by: Koni Marti <koni.marti@gmail.com>