ipc: wait for app to be ready before accepting commands
Pass a cancellable context to the IPC server and defer accepting
connections until the first worker message has been received by the main
thread.

This is not a real fix for the :split command not working at startup.
:split requires a mail store to be present on the currently selected
account tab+folder. I could have waited for a types.DirectoryContents
message but it seems silly and bug prone since such a message may never
arrive and we don't want to pool IPC messages forever.

Also, a types.DirectoryContents message may arrive for one account but
:split is only effective *per account* and the default selected account
tab is the first one.

In any case, this band aid prevents aerc from crashing or breaking down
the terminal when running IPC commands in the aerc-startup hook.

Fixes: https://lists.sr.ht/~rjarry/aerc-devel/%3CCYPN7AVYQ69S.WV0T67VM6WX3%40wegtam.com%3E
Fixes: https://todo.sr.ht/~rjarry/aerc/173
Reported-by: Jens Grassel <jens@wegtam.com>
Reported-by: Justine Smithies <justine@smithies.me.uk>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Thomas Böhler <witcher@wiredspace.de>
Reviewed-by: Thomas Böhler <witcher@wiredspace.de>